add capybara, rework spec support
