+++ Gemfile
@@ -5,11 +5,13 @@ source "https://rubygems.org"
 gemspec
 
 gem "appraisal"
+gem "capybara"
 gem "rack-session"
 gem "rack-test"
 gem "rake"
 gem "rspec"
 gem "rubocop"
+gem "rubocop-capybara"
 gem "rubocop-performance"
 gem "rubocop-rake"
 gem "rubocop-rspec"
