+++ lib/omniauth/passthrough.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 require_relative "passthrough/action_dispatch_helper"
+require_relative "passthrough/capybara_helper"
+require_relative "passthrough/capybara_middleware"
 require_relative "passthrough/rack_helper"
 require_relative "passthrough/version"
 require_relative "strategies/passthrough"
