+++ lib/omniauth/passthrough/capybara_helper.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module OmniAuth
+  module Passthrough
+    module CapybaraHelper
+      def omniauth_passthrough(params = {})
+      end
+    end
+  end
+end
