+++ lib/omniauth/passthrough/capybara_middleware.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module OmniAuth
+  module Passthrough
+    class CapybaraMiddleware
+    end
+  end
+end
