+++ spec/lib/omniauth/passthrough/capybara_helper_spec.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require "omniauth/passthrough/capybara_helper"
+
+RSpec.describe OmniAuth::Passthrough::CapybaraHelper do
+  include described_class
+
+  describe "#omniauth_passthrough", :capybara do
+  end
+end
