+++ spec/lib/omniauth/passthrough/capybara_middleware_spec.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require "omniauth/passthrough/capybara_middleware"
+
+RSpec.describe OmniAuth::Passthrough::CapybaraMiddleware do
+end
