+++ spec/lib/omniauth/passthrough_spec.rb
